Backend/requirements.txt
asgiref==3.8.1
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.1.8
colorama==0.4.6
Django==5.1.6
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
Flask==3.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
PyJWT==2.12.1
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.6.3
Werkzeug==3.1.3
